rename component to transport
