+++ core.go
@@ -7,13 +7,13 @@ import (
 	"log"
 	"time"
 
-	"github.com/xxf098/lite-proxy/component/resolver"
 	"github.com/xxf098/lite-proxy/config"
 	_ "github.com/xxf098/lite-proxy/config"
 	"github.com/xxf098/lite-proxy/dns"
 	"github.com/xxf098/lite-proxy/outbound"
 	"github.com/xxf098/lite-proxy/proxy"
 	"github.com/xxf098/lite-proxy/request"
+	"github.com/xxf098/lite-proxy/transport/resolver"
 	"github.com/xxf098/lite-proxy/tunnel"
 	"github.com/xxf098/lite-proxy/tunnel/adapter"
 	"github.com/xxf098/lite-proxy/tunnel/http"
