+++ dns/resolver.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/xxf098/lite-proxy/common/cache"
 	"github.com/xxf098/lite-proxy/common/picker"
-	"github.com/xxf098/lite-proxy/component/resolver"
+	"github.com/xxf098/lite-proxy/transport/resolver"
 	"golang.org/x/sync/singleflight"
 
 	// "github.com/Dreamacro/clash/component/trie"
