+++ outbound/shadowsocksr.go
@@ -10,10 +10,10 @@ import (
 	"github.com/Dreamacro/go-shadowsocks2/core"
 	"github.com/Dreamacro/go-shadowsocks2/shadowaead"
 	"github.com/Dreamacro/go-shadowsocks2/shadowstream"
-	"github.com/xxf098/lite-proxy/component/dialer"
-	"github.com/xxf098/lite-proxy/component/ssr/obfs"
-	"github.com/xxf098/lite-proxy/component/ssr/protocol"
 	C "github.com/xxf098/lite-proxy/constant"
+	"github.com/xxf098/lite-proxy/transport/dialer"
+	"github.com/xxf098/lite-proxy/transport/ssr/obfs"
+	"github.com/xxf098/lite-proxy/transport/ssr/protocol"
 )
 
 type ShadowSocksR struct {
