+++ transport/ssr/obfs/http_simple.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 
 	"github.com/xxf098/lite-proxy/common/pool"
-	"github.com/xxf098/lite-proxy/component/ssr/tools"
+	"github.com/xxf098/lite-proxy/transport/ssr/tools"
 )
 
 func init() {
