+++ transport/ssr/obfs/tls1.2_ticket_auth.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	"github.com/xxf098/lite-proxy/common/pool"
-	"github.com/xxf098/lite-proxy/component/ssr/tools"
+	"github.com/xxf098/lite-proxy/transport/ssr/tools"
 )
 
 func init() {
