+++ transport/ssr/protocol/auth_aes128_md5.go
@@ -1,6 +1,6 @@
 package protocol
 
-import "github.com/xxf098/lite-proxy/component/ssr/tools"
+import "github.com/xxf098/lite-proxy/transport/ssr/tools"
 
 func init() {
 	register("auth_aes128_md5", newAuthAES128MD5, 9)
