+++ transport/ssr/protocol/auth_sha1_v4.go
@@ -9,7 +9,7 @@ import (
 	"net"
 
 	"github.com/xxf098/lite-proxy/common/pool"
-	"github.com/xxf098/lite-proxy/component/ssr/tools"
+	"github.com/xxf098/lite-proxy/transport/ssr/tools"
 )
 
 func init() {
