+++ transport/ssr/protocol/base.go
