+++ transport/ssr/protocol/packet.go
@@ -4,7 +4,7 @@ import (
 	"bytes"
 	"net"
 
-	"github.com/xxf098/lite-proxy/component/ssr/tools"
+	"github.com/xxf098/lite-proxy/transport/ssr/tools"
 )
 
 type PacketConn struct {
