+++ transport/ssr/protocol/stream.go
@@ -5,7 +5,7 @@ import (
 	"net"
 
 	"github.com/xxf098/lite-proxy/common/pool"
-	"github.com/xxf098/lite-proxy/component/ssr/tools"
+	"github.com/xxf098/lite-proxy/transport/ssr/tools"
 )
 
 type Conn struct {
