+++ transport/ssr/tools/crypto.go
