+++ transport/ssr/tools/random.go
